+++ vulture_whitelist.py
@@ -32,7 +32,7 @@ run_bc_check_agent
 reset_for_tests
 run_completeness_check_agent
 run_config_sync_agent
-run_security_posture_agent
+
 _absorb_summary_typos
 best_k
 failure_summary
@@ -123,8 +123,6 @@ frontend_sync_periodic
 frontend_sync_interval_seconds
 member_sync_interval_seconds
 meta_periodic
-security_posture_periodic
-security_posture_interval_seconds
 run_health_periodic
 run_health_memory_path
 diagnostic_periodic
@@ -259,7 +257,6 @@ run_bc_check_pass
 run_completeness_check_pass
 run_config_sync_pass
 run_copy_paste_pass
-run_security_posture_pass
 run_data_dir_gc_pass
 dir_size_bytes
 oversized_items